Title: 정수 사이의 배수, Time: 8ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 251023/정수 사이의 배수/multiplication-between-numbers.cpp
@@ -7,7 +7,7 @@ int main()
 {
     // Please write your code here.
     int a, b, fs_sum = 0, fs_cnts = 0;
-    float fs_avg = 0.0;
+    double fs_avg = 0.0;
     cin >> a >> b;
 
     for(int i = a; i <= b; i++)
